+++ .env.example
@@ -1,18 +1,20 @@
 SERVER_SQLSERVER=
 DB_SQLSERVER=
 HOST_SERVER=
-FILE_ESFERAS=
-FILE_MUELLE=
-FILE_DUCTO=
+
 DATABASE_DB=
 HOST_DB=
 USER_DB=
 PASSWORD_DB=
 PORT_DB=
-FOLDER_MES_ESFERAS=
-FOLDER_MES_MUELLE=
-FOLDER_MES_DUCTO=
+
 DATABASE_DB_DOC=
 HOST_DB_DOC=
 USER_DB_DOC=
 PASSWORD_DB_DOC=
+
+DATABASE_DB_DUCA=
+HOST_DB_DUCA=
+USER_DB_DUCA=
+PASSWORD_DB_DUCA=
+PORT_DB_DUCA=
